refactor: 거리 정보를 담지 않은 음식점 DTO 네이밍을 RestaurantSearchWithoutDistanceQueryResponse로 변경 (#583)
