+++ backend/src/main/java/com/celuveat/restaurant/presentation/MainPageController.java
@@ -4,7 +4,7 @@ import com.celuveat.common.PageResponse;
 import com.celuveat.common.auth.LooseAuth;
 import com.celuveat.restaurant.presentation.dto.RegionCodeCondRequest;
 import com.celuveat.restaurant.query.RestaurantQueryService;
-import com.celuveat.restaurant.query.dto.RestaurantByRegionCodeQueryResponse;
+import com.celuveat.restaurant.query.dto.RestaurantSearchWithoutDistanceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -21,12 +21,12 @@ public class MainPageController {
     private final RestaurantQueryService restaurantQueryService;
 
     @GetMapping("/address")
-    ResponseEntity<PageResponse<RestaurantByRegionCodeQueryResponse>> findAllByRegionCode(
+    ResponseEntity<PageResponse<RestaurantSearchWithoutDistanceResponse>> findAllByRegionCode(
             @LooseAuth Long memberId,
             @ModelAttribute RegionCodeCondRequest regionCodeCondRequest,
             @PageableDefault(size = 18) Pageable pageable
     ) {
-        Page<RestaurantByRegionCodeQueryResponse> result = restaurantQueryService.findAllByRegionCode(
+        Page<RestaurantSearchWithoutDistanceResponse> result = restaurantQueryService.findAllByRegionCode(
                 regionCodeCondRequest.toCondition(), pageable, memberId
         );
         return ResponseEntity.ok(PageResponse.from(result));
