+++ backend/src/main/java/com/celuveat/restaurant/presentation/dto/RegionCodeCondRequest.java
@@ -1,6 +1,6 @@
 package com.celuveat.restaurant.presentation.dto;
 
-import com.celuveat.restaurant.query.dao.RestaurantByRegionCodeQueryResponseDao.RegionCodeCond;
+import com.celuveat.restaurant.query.dao.RestaurantSearchWithoutDistanceQueryResponseDao.RegionCodeCond;
 import java.util.List;
 
 public record RegionCodeCondRequest(
