+++ backend/src/main/java/com/celuveat/restaurant/query/RestaurantQueryService.java
@@ -3,15 +3,15 @@ package com.celuveat.restaurant.query;
 import static com.celuveat.restaurant.query.mapper.RestaurantRelocator.relocateCelebDataFirstByCelebId;
 import static com.celuveat.restaurant.query.mapper.RestaurantRelocator.relocateCelebDataFirstInResponsesByCelebId;
 
-import com.celuveat.restaurant.query.dao.RestaurantByRegionCodeQueryResponseDao;
-import com.celuveat.restaurant.query.dao.RestaurantByRegionCodeQueryResponseDao.RegionCodeCond;
+import com.celuveat.restaurant.query.dao.RestaurantSearchWithoutDistanceQueryResponseDao;
+import com.celuveat.restaurant.query.dao.RestaurantSearchWithoutDistanceQueryResponseDao.RegionCodeCond;
 import com.celuveat.restaurant.query.dao.RestaurantDetailQueryResponseDao;
 import com.celuveat.restaurant.query.dao.LikedRestaurantQueryResponseDao;
 import com.celuveat.restaurant.query.dao.RestaurantSearchQueryResponseDao;
 import com.celuveat.restaurant.query.dao.RestaurantSearchQueryResponseDao.LocationSearchCond;
 import com.celuveat.restaurant.query.dao.RestaurantSearchQueryResponseDao.RestaurantSearchCond;
 import com.celuveat.restaurant.query.dto.LikedRestaurantQueryResponse;
-import com.celuveat.restaurant.query.dto.RestaurantByRegionCodeQueryResponse;
+import com.celuveat.restaurant.query.dto.RestaurantSearchWithoutDistanceResponse;
 import com.celuveat.restaurant.query.dto.RestaurantDetailQueryResponse;
 import com.celuveat.restaurant.query.dto.RestaurantSearchQueryResponse;
 import jakarta.annotation.Nullable;
@@ -30,7 +30,7 @@ public class RestaurantQueryService {
     private final RestaurantDetailQueryResponseDao restaurantDetailQueryResponseDao;
     private final RestaurantSearchQueryResponseDao restaurantSearchQueryResponseDao;
     private final LikedRestaurantQueryResponseDao likedRestaurantQueryResponseDao;
-    private final RestaurantByRegionCodeQueryResponseDao restaurantByRegionCodeQueryResponseDao;
+    private final RestaurantSearchWithoutDistanceQueryResponseDao restaurantSearchWithoutDistanceQueryResponseDao;
 
     public RestaurantDetailQueryResponse findRestaurantDetailById(
             Long restaurantId,
@@ -58,12 +58,12 @@ public class RestaurantQueryService {
         return relocateCelebDataFirstInResponsesByCelebId(celebId, response);
     }
 
-    public Page<RestaurantByRegionCodeQueryResponse> findAllByRegionCode(
+    public Page<RestaurantSearchWithoutDistanceResponse> findAllByRegionCode(
             RegionCodeCond cond,
             Pageable pageable,
             @Nullable Long memberId
     ) {
-        return restaurantByRegionCodeQueryResponseDao.find(cond, pageable, memberId);
+        return restaurantSearchWithoutDistanceQueryResponseDao.findByRegionCode(cond, pageable, memberId);
     }
 
     public Page<RestaurantSearchQueryResponse> findAllNearByDistanceWithoutSpecificRestaurant(
