+++ backend/src/main/java/com/celuveat/restaurant/query/dto/LikedRestaurantQueryResponse.java
@@ -4,6 +4,7 @@ import com.celuveat.restaurant.command.domain.Restaurant;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
+// TODO 여기에 조회수, 평점 등에 대한 정보가 없는데 이대로 갈건지?, 만약 변경된다면 해당 클래스 대신 RestaurantWithoutDistanceQueryResponse를 사용할 것
 public record LikedRestaurantQueryResponse(
         Long id,
         String name,
